refactor: Transforma parsers em helper
