+++ src/helpers/translateAmount.ts
@@ -1,0 +1,6 @@
+import { ITransaction } from "@/modules/home/interfaces/transaction";
+import { convertToBrazilianHelper } from './convertNumber';
+
+export const translateAmount = ({ amount }: ITransaction): string => {
+  return new convertToBrazilianHelper().convert(amount);
+};
